Move installable kit types to /models
