+++ src/client.spec.ts
@@ -11,7 +11,6 @@ import { attempt } from 'lodash';
 import { GravwellClient } from './client';
 import {
 	AutoExtractorsFilter,
-	BuildableKit,
 	CreatableBroadcastNotification,
 	CreatableMacro,
 	CreatablePlaybook,
@@ -24,9 +23,7 @@ import {
 	CreatableUser,
 	DashboardsFilter,
 	GetAPIVersionResponse,
-	InstallableKit,
 	IsValidAutoExtractorSyntaxResponse,
-	KitInstallationStatus,
 	MacrosFilter,
 	MyNotificationsMessageReceived,
 	MyNotificationsMessageSent,
@@ -56,6 +53,7 @@ import {
 	Actionable,
 	AutoExtractor,
 	AutoExtractorModule,
+	BuildableKit,
 	CreatableActionable,
 	CreatableAutoExtractor,
 	CreatableDashboard,
@@ -66,6 +64,8 @@ import {
 	Dashboard,
 	FileMetadata,
 	Group,
+	InstallableKit,
+	KitInstallationStatus,
 	LocalKit,
 	LogLevel,
 	Macro,
