+++ src/models/kit/index.ts
@@ -7,13 +7,19 @@
  **************************************************************************/
 
 export * from './buildable-kit';
+export * from './installable-kit';
 export * from './is-local-kit';
 export * from './is-remote-kit';
+export * from './kit-installation-status';
 export * from './local-kit';
 export * from './raw-buildable-kit';
+export * from './raw-installable-kit';
+export * from './raw-kit-installation-status';
 export * from './raw-local-kit';
 export * from './raw-remote-kit';
 export * from './remote-kit';
+export * from './to-kit-installation-status';
 export * from './to-local-kit';
 export * from './to-raw-buildable-kit';
+export * from './to-raw-installable-kit';
 export * from './to-remote-kit';
