+++ src/models/kit/installable-kit.ts
@@ -1,0 +1,25 @@
+/*************************************************************************
+ * Copyright 2020 Gravwell, Inc. All rights reserved.
+ * Contact: <legal@gravwell.io>
+ *
+ * This software may be modified and distributed under the terms of the
+ * MIT license. See the LICENSE file for details.
+ **************************************************************************/
+
+export interface InstallableKit {
+	labels?: Array<string>;
+	isGlobal?: boolean;
+
+	overwriteExisting?: boolean;
+	allowUnsigned?: boolean;
+	allowExternalResource?: boolean;
+
+	settings?: Array<{
+		type: 'macro value';
+		name: string;
+		description: string;
+		defaultValue: string;
+		value: string | null;
+		valueType: 'tag' | 'string';
+	}>;
+}
